print

+++ roi_model.py
@@ -460,7 +460,7 @@ class Encoder(nn.Module):
         LongTensor = torch.cuda.LongTensor
         ByteTensor = torch.cuda.ByteTensor
         return (torch.zeros(self.num_layers, batch_size, self.hidden_dim),
-                torch.zeros(self.num_layers, batch_size, self.hidden_dim)).type(FloatTensor)
+                torch.zeros(self.num_layers, batch_size, self.hidden_dim))
 
     def forward(self, inputs):
         # Push through RNN layer (the ouput is irrelevant)
